+++ proto-lens-protoc/src/Data/ProtoLens/Compiler/Definitions.hs
@@ -70,6 +70,7 @@ import Proto.Google.Protobuf.Descriptor
     , FieldDescriptorProto'Type(..)
     , FileDescriptorProto
     , MethodDescriptorProto
+    , MethodOptions
     , ServiceDescriptorProto
     )
 import Proto.Google.Protobuf.Descriptor_Fields
@@ -160,6 +161,7 @@ data MethodInfo = MethodInfo
     , methodOutput :: Text
     , methodClientStreaming :: Bool
     , methodServerStreaming :: Bool
+    , methodOptions :: MethodOptions
     }
 
 -- | Information about a single field of a proto message,
@@ -352,6 +354,7 @@ collectServices fd = fmap (toServiceInfo $ fd ^. package) $ fd ^. service
             , methodOutput = fromString . T.unpack $ md ^. outputType
             , methodClientStreaming = md ^. clientStreaming
             , methodServerStreaming = md ^. serverStreaming
+            , methodOptions = md ^. options
             }
 
 messageAndEnumDefs ::